requirements.txt
# Kern
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
python-dotenv==1.*
requests==2.*


# PDF/Parsing
pypdf==5.*


# Embeddings
sentence-transformers==3.*
transformers==4.*
torch>=2.2


# Vector Store
faiss-cpu>=1.8.0


# Optional: Reranking (Cross‑Encoder)
# pip install -U sentence-transformers && ein BGE Reranker‑Modell manuell laden